Split submoduels
